Add files via upload
Prodigy_Data-Science_Task-01
Creating a Bar or Histogram Visualization You can use this code to create a bar or histogram visualization of the distribution of a categorical or continuous variable in your population dataset. For example, you can visualize the distribution of ages or gender in the population.

Population Age Distribution Visualization
This project allows you to create a bar chart to visualize the age distribution within a population dataset. It's a tool for gaining insights into the age demographics of a population.